feat: added a payed subscription system
This will be used to track user subscriptions, I don't know if I will keep using the Stripe systems, because they are not known to be the best system.
I'm searching for open source alternative for go, Lago seems cool but seems to have some issues in the documentation.
I also don't want to handle it myself, as I don't want to manage users money because I'm not the best programmer in the world.